+++ src/domain/core/src/services/ingest/polling_ingest_service.rs
@@ -24,6 +24,12 @@ use crate::*;
 
 #[async_trait::async_trait]
 pub trait PollingIngestService: Send + Sync {
+    /// Returns an active polling source, if any
+    async fn get_active_polling_source(
+        &self,
+        dataset_ref: &DatasetRef,
+    ) -> Result<Option<(Multihash, MetadataBlockTyped<SetPollingSource>)>, GetDatasetError>;
+
     /// Uses polling source definition in metadata to ingest data from an
     /// external source
     async fn ingest(