+++ src/domain/core/src/services/ingest/push_ingest_service.rs
@@ -21,6 +21,12 @@ use crate::*;
 
 #[async_trait::async_trait]
 pub trait PushIngestService: Send + Sync {
+    /// Returns the set of active push sources
+    async fn get_active_push_sources(
+        &self,
+        dataset_ref: &DatasetRef,
+    ) -> Result<Vec<(Multihash, MetadataBlockTyped<AddPushSource>)>, GetDatasetError>;
+
     /// Uses push source defenition in metadata to ingest data from the
     /// specified source.
     ///