+++ src/infra/core/src/ingest/polling_ingest_service_impl.rs
@@ -688,6 +688,25 @@ impl PollingIngestServiceImpl {
 #[async_trait::async_trait]
 impl PollingIngestService for PollingIngestServiceImpl {
     #[tracing::instrument(level = "info", skip_all, fields(%dataset_ref))]
+    async fn get_active_polling_source(
+        &self,
+        dataset_ref: &DatasetRef,
+    ) -> Result<Option<(Multihash, MetadataBlockTyped<SetPollingSource>)>, GetDatasetError> {
+        let dataset = self.dataset_repo.get_dataset(dataset_ref).await?;
+
+        // TODO: Support source evolution
+        let source = dataset
+            .as_metadata_chain()
+            .iter_blocks()
+            .filter_map_ok(|(h, b)| b.into_typed::<SetPollingSource>().map(|b| (h, b)))
+            .try_first()
+            .await
+            .int_err()?;
+
+        Ok(source)
+    }
+
+    #[tracing::instrument(level = "info", skip_all, fields(%dataset_ref))]
     async fn ingest(
         &self,
         dataset_ref: &DatasetRef,