+++ src/infra/core/src/ingest/push_ingest_service_impl.rs
@@ -316,6 +316,23 @@ impl PushIngestServiceImpl {
 
 #[async_trait::async_trait]
 impl PushIngestService for PushIngestServiceImpl {
+    #[tracing::instrument(level = "info", skip_all, fields(%dataset_ref))]
+    async fn get_active_push_sources(
+        &self,
+        dataset_ref: &DatasetRef,
+    ) -> Result<Vec<(Multihash, MetadataBlockTyped<AddPushSource>)>, GetDatasetError> {
+        use futures::TryStreamExt;
+
+        // TODO: Support source disabling and evolution
+        let dataset = self.dataset_repo.get_dataset(dataset_ref).await?;
+        let stream = dataset
+            .as_metadata_chain()
+            .iter_blocks()
+            .filter_map_ok(|(h, b)| b.into_typed().map(|b| (h, b)));
+
+        Ok(stream.try_collect().await.int_err()?)
+    }
+
     #[tracing::instrument(level = "info", skip_all, fields(%dataset_ref, %url, ?media_type))]
     async fn ingest_from_url(
         &self,